began refactoring ui and added mentioned the CC0 in the publish dialog
